mindpilot/requirements.txt
openai>=1.0.0
requests>=2.31.0
feedparser>=6.0.10
arxiv>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0
scipy>=1.11.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
networkx>=3.1
python-dotenv>=1.0.0
jinja2>=3.1.2
markdown>=3.4.4
rouge-score>=0.1.2
nltk>=3.8.1
aiohttp>=3.9.0
asyncio-pool>=0.6.0
pydantic>=2.0.0
rich>=13.5.0
typer>=0.9.0
python-docx>=1.1.0
